requirements.txt
django==2.2.16
django-simpleui
django-cors-headers
django-debug-toolbar
djangorestframework
django-rest-auth
django-allauth
# mysqlclient 
pymysql
Pillow
django-storages
boto3
django-ckeditor
djangorestframework_simplejwt
djoser
django-s3direct
aliyun-python-sdk-sts
oss2
django-parsley
httplib2
django-mysql
braintree
gunicorn
drf-yasg
qrcode
gevent
django-aliyun-oss2-storage
django-filer
django-mptt
easy_thumbnails
django-polymorphic
django-ckeditor-filebrowser-filer
django-crequest
pyyaml
jsonpickle
pytz
tzlocal
wheel
django-redis